refactor: rearrange roles of view and controller
